+++ .env.test
@@ -1,0 +1,2 @@
+JWT_SECRET="test-secret"
+DATABASE_URL="postgresql://postgres:1111@db:5432/test?schema=public"
