+++ __tests__/setup.ts
@@ -1,0 +1,2 @@
+import next from "next";
+next({})
