+++ package.json
@@ -1,49 +1,61 @@
 {
-  "name": "coffeemtr",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint",
-    "pdev": "prisma migrate dev"
-  },
-  "dependencies": {
-    "@next/font": "13.1.2",
-    "@prisma/client": "^4.9.0",
-    "@types/js-cookie": "^3.0.3",
-    "@types/node": "18.11.18",
-    "@types/react": "18.0.26",
-    "@types/react-big-calendar": "^1.6.3",
-    "@types/react-day-picker": "^5.3.0",
-    "@types/react-dom": "18.0.10",
-    "@types/styled-components": "^5.1.26",
-    "babel-plugin-styled-components": "^2.0.7",
-    "bcryptjs": "^2.4.3",
-    "bootstrap": "^5.2.3",
-    "chart.js": "^4.2.0",
-    "date-fns": "^2.29.3",
-    "dayjs": "^1.11.7",
-    "eslint": "8.32.0",
-    "eslint-config-next": "13.1.2",
-    "js-cookie": "^3.0.1",
-    "jsonwebtoken": "^9.0.0",
-    "next": "13.1.2",
-    "prisma": "^4.9.0",
-    "react": "18.2.0",
-    "react-big-calendar": "^1.6.9",
-    "react-bootstrap": "^2.7.0",
-    "react-calendar": "^4.0.0",
-    "react-chartjs-2": "^5.2.0",
-    "react-datepicker": "^4.11.0",
-    "react-day-picker": "^8.7.0",
-    "react-dom": "18.2.0",
-    "styled-components": "^5.3.6",
-    "swr": "^2.0.0",
-    "typescript": "4.9.4"
-  },
-  "devDependencies": {
-    "@types/react-calendar": "^3.9.0"
-  }
+	"name": "coffeemtr",
+	"version": "0.1.0",
+	"private": true,
+	"scripts": {
+		"dev": "next dev",
+		"build": "next build",
+		"start": "next start",
+		"lint": "next lint",
+		"pdev": "prisma migrate dev",
+		"test": "jest"
+	},
+	"dependencies": {
+		"@next/font": "13.1.2",
+		"@prisma/client": "^4.9.0",
+		"@types/bcryptjs": "^2.4.2",
+		"@types/js-cookie": "^3.0.3",
+		"@types/jsonwebtoken": "^9.0.1",
+		"@types/node": "18.11.18",
+		"@types/react": "18.0.26",
+		"@types/react-big-calendar": "^1.6.3",
+		"@types/react-day-picker": "^5.3.0",
+		"@types/react-dom": "18.0.10",
+		"@types/styled-components": "^5.1.26",
+		"@typescript-eslint/parser": "^5.59.0",
+		"babel-plugin-styled-components": "^2.0.7",
+		"bcryptjs": "^2.4.3",
+		"bootstrap": "^5.2.3",
+		"chart.js": "^4.2.0",
+		"date-fns": "^2.29.3",
+		"dayjs": "^1.11.7",
+		"eslint": "8.32.0",
+		"eslint-config-next": "13.1.2",
+		"js-cookie": "^3.0.1",
+		"jsonwebtoken": "^9.0.0",
+		"next": "13.1.2",
+		"prisma": "^4.9.0",
+		"react": "18.2.0",
+		"react-big-calendar": "^1.6.9",
+		"react-bootstrap": "^2.7.0",
+		"react-calendar": "^4.0.0",
+		"react-chartjs-2": "^5.2.0",
+		"react-datepicker": "^4.11.0",
+		"react-day-picker": "^8.7.0",
+		"react-dom": "18.2.0",
+		"styled-components": "^5.3.6",
+		"swr": "^2.0.0",
+		"typescript": "4.9.4"
+	},
+	"devDependencies": {
+		"@types/jest": "^29.5.1",
+		"@types/react-calendar": "^3.9.0",
+		"@types/react-datepicker": "^4.10.0",
+		"@types/supertest": "^2.0.12",
+		"jest": "^29.5.0",
+		"node-mocks-http": "^1.12.2",
+		"supertest": "^6.3.3",
+		"ts-jest": "^29.1.0",
+		"ts-node": "^10.9.1"
+	}
 }
