+++ prisma/schema.prisma
@@ -15,6 +15,7 @@ model User {
   id          Int      @id @default(autoincrement()) @map("user_id")
   email       String   @unique
   password    String
+  token       String @default("")
   drinks      DrinkCoffeeHistory[]
   createdAt   DateTime @default(now()) @map(name: "created_at")
   @@map("users")
@@ -23,6 +24,7 @@ model User {
 model DrinkCoffeeHistory {
   id                Int      @id @default(autoincrement()) @map("drink_coffee_history_id")
   caffeine_contents_mg Int
+  drankdAt         DateTime @default(now()) @map(name: "drank_at")
   createdAt         DateTime @default(now()) @map(name: "created_at")
   user              User?    @relation(fields: [userId], references: [id])
   userId            Int?
