+++ src/pages/api/drink/calendar.ts
@@ -1,5 +1,6 @@
 import { NextApiRequest, NextApiResponse } from "next";
-import { PrismaClient, Prisma } from "@prisma/client";
+import { PrismaClient, User } from "@prisma/client";
+import { authUser } from "@/lib/lib";
 import { genInputDate } from "@/const";
 
 const prisma = new PrismaClient();
@@ -8,9 +9,13 @@ export default async function handler(
   req: NextApiRequest,
   res: NextApiResponse
 ) {
+  const user: User | null = await authUser(req.headers.authorization);
+  if (!user) {
+    res.status(401).json({ msg: "unauthorized." });
+    return;
+  }
   if (req.method === "GET") {
     let year_month = req.query.ym;
-	console.log(year_month)
     const { gte, lte, yms } = genInputDate(year_month);
     year_month = yms;
     const histList = await prisma.drinkCoffeeHistory.findMany({
